refactor FailToSend to PartialFailureToSend
