+++ src/script/components/MessagesList/Message/ContentMessage/Warnings/PartialFailureToSend/index.ts
@@ -17,4 +17,4 @@
  *
  */
 
-export * from './FailedToSend';
+export * from './PartialFailureToSend';
